LPS-66853 Move logic to retrieve properties to SourceFormatter. We only want to do this once before we start threading. If every sourceProcessor starts scanning the directories for source-formatter.properties files at the same time, things will slow down a lot
